Refactoring files
Moved every function to their own file and refactored loading them in index.html using jQuery.
