+++ scripts/meta/removeUnNeededSpace.js
@@ -1,0 +1,5 @@
+function removeUnNeededSpace(input) {
+	input = input.replace(/\*|\r?\n|\r|\s+/g, " ");
+	input = input.replace(/;;/g, ";");
+	return input.replace(/\s?-\s?/g, "-").trim();
+}
