+++ scripts/meta/replaceEmptyRule.js
@@ -1,0 +1,4 @@
+function replaceEmptyRule(input)
+{ 
+	return input.replace(/:;/g,":"); 
+}
