+++ scripts/month/monthsAddSpace.js
@@ -1,0 +1,5 @@
+function monthsAddSpace(input) {
+	return input.replace(/;(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)/g, (_1, _2) => {
+		return "; " + _2;
+	});
+}
